requirements.txt
openai
faiss-cpu
tqdm
pandas
numpy
python-dotenv
transformers
torch